backend/requirements.txt
fastapi
supabase
python-dotenv
uvicorn
uvicorn
python-multipart # For file uploads 